Delete old redfish subscriptions
When hardware is moved the old subscriptions the bmc can have
subscriptions for the old, and now wrong xname.  With these
changes hmcollector will automatically delete these subscriptions
when it finds them.

Here are the cases where it can find a bad subscirption
- At start up it will check all subscriptions.
- Periodically it will check for it's own subscriptions, and it will
delete any bad ones that it finds, however, in this case it does not
necessarily read all the subscritpions. It only reads until it finds
the subscription it is looking for.

CASMHMS-5685