Delete old redfish subscriptions (#49)
When hardware is moved the bmc may retain its old subscriptions.
These old subscriptions refer to the old location/xname. This causes
problems when the hardware sends events. The listeners mistakenly
think the events are for the old location/xname.

This commit changes hmcollector to automatically delete these
old subscriptions.

Here are the cases where it can find a bad subscription
- At start up it will check all subscriptions.
- Periodically it will check for its own subscriptions, and it will
delete any bad ones that it finds, however, in this case it does not
necessarily read all the subscriptions. It only reads until it finds
the subscription it is looking for.

CASMHMS-5685